backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
pymysql
python-jose[cryptography]
passlib[bcrypt] 
python-multipart
pydantic[email]
openpyxl
pandas